requirements.txt
pandas==1.3.2
Flask==2.2.2
seaborn==0.11.2
matplotlib==3.4.3
numpy==1.19.0
scikit_learn==0.24.2
mysql-connector-python==8.0.28
Flask-Cors==3.0.10